refactor(auth): extract requireTrainerOwnsSession to ownership.ts
Moves the duplicated requireTrainerOwnsSession helper from audio.ts and
aiInsights.ts into a shared plain module src/lib/auth/ownership.ts.
No "use server" directive — keeps it a regular importable module.

Closes #123
